+++ messaging/src/main/java/com/crio/messaging/message/MessageType.java
@@ -1,4 +1,4 @@
-package com.crio.messaging;
+package com.crio.messaging.message;
 
 public enum MessageType {
     TEXT,
